add remote code

+++ src/firmware/drivers/eMPL/inv_mpu_dmp_motion_driver.h
@@ -93,7 +93,7 @@ int dmp_enable_gyro_cal(unsigned char enable);
 int dmp_read_fifo(short *gyro, short *accel, long *quat,
     unsigned long *timestamp, short *sensors, unsigned char *more);
 
-int dmp_read_quat(long *quat, unsigned char *more);
+int dmp_read_quat(float *quat, unsigned char *more);
 
 #endif  /* #ifndef _INV_MPU_DMP_MOTION_DRIVER_H_ */
